feat: refactor `useTeamTasks` into specialized hooks and add `getUserDisplayName` helper to fix user name display.
