+++ apps/web/core/lib/helpers/index.ts
@@ -11,4 +11,3 @@ export * from './strings';
 export * from './plan-day-badge';
 export * from './drag-and-drop';
 export * from './type-guards';
-export * from '../helpers';
